// frontend/src/types/prisma/types.ts
// This file was auto-generated by prisma-generator-typescript-interfaces

export interface User {
	id: string;
	email: string;
	password: string;
	name: string | null;
	createdAt: Date;
	updatedAt: Date;
	bots?: Bot[];
	servers?: Server[];
	sessions?: Session[];
}

export interface Session {
	id: string;
	userId: string;
	token: string;
	expiresAt: Date;
	createdAt: Date;
	user?: User;
}

export interface Server {
	id: string;
	name: string;
	description: string | null;
	ownerId: string;
	createdAt: Date;
	updatedAt: Date;
	serverid: string | null;
	bots?: Bot[];
	owner?: User;
}

export interface Bot {
	id: string;
	name: string;
	token: string;
	active: boolean;
	ownerId: string;
	prefix: string;
	createdAt: Date;
	updatedAt: Date;
	description: string;
	tags: string[];
	icon: string | null;
	serverId: string;
	analytics?: Analytics[];
	owner?: User;
	server?: Server;
	configurations?: BotConfig[];
	commands?: Command[];
	events?: Event[];
}

export interface Command {
	id: string;
	name: string;
	description: string;
	response: string;
	enabled: boolean;
	botId: string;
	createdAt: Date;
	updatedAt: Date;
	bot?: Bot;
	permissions?: CommandPermission[];
	usage?: CommandUsage[];
}

export interface CommandPermission {
	id: string;
	commandId: string;
	roleId: string;
	allow: boolean;
	createdAt: Date;
	command?: Command;
}

export interface Event {
	id: string;
	name: string;
	action: string;
	enabled: boolean;
	botId: string;
	createdAt: Date;
	updatedAt: Date;
	bot?: Bot;
}

export interface BotConfig {
	id: string;
	key: string;
	value: string;
	botId: string;
	createdAt: Date;
	updatedAt: Date;
	bot?: Bot;
}

export interface CommandUsage {
	id: string;
	commandId: string;
	userId: string;
	guildId: string;
	channelId: string;
	timestamp: Date;
	command?: Command;
}

export interface Analytics {
	id: string;
	botId: string;
	eventType: string;
	data: JsonValue;
	timestamp: Date;
	bot?: Bot;
}

type JsonValue = string | number | boolean | { [key in string]?: JsonValue } | Array<JsonValue> | null;